requirements.txt
fasthtml
ollama
huggingface_hub
python-dotenv 